Adding device ops handling

+++ tools/profiler/profiler.py
@@ -13,7 +13,7 @@ from tracy.process_ops_logs import process_ops
 
 
 @contextmanager
-def trace(log_dir: str, port: int):
+def trace(log_dir: str, port: int, host_only: bool = False):
     os.makedirs(log_dir, exist_ok=True)
 
     TT_METAL_RUNTIME_ROOT = os.environ.get(
@@ -28,8 +28,16 @@ def trace(log_dir: str, port: int):
     profiler_csv_file_path = (
         TT_METAL_RUNTIME_ROOT + "/generated/profiler/reports/ops_perf_results.csv"
     )
+    os.environ["TRACY_PORT"] = str(port)
+    if not host_only:
+        os.environ["TT_METAL_CLEAR_L1"] = "1"
+        os.environ["TT_METAL_DEVICE_PROFILER"] = "1"
+        os.environ["TTNN_OP_PROFILER"] = "1"
+        os.environ["TT_METAL_DEVICE_PROFILER_DISPATCH"] = "0"
+        os.environ["TT_METAL_PROFILER_CPP_POST_PROCESS"] = "1"
 
-    shutil.rmtree(profiler_logs_dir)
+    if os.path.exists(profiler_logs_dir):
+        shutil.rmtree(profiler_logs_dir)
     os.makedirs(profiler_logs_dir)
 
     tracy_capture_tool_command = (
